pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a-sdlc"
version = "0.1.0"
description = "SDLC Automation System for Claude Code - BrainGrid-style documentation and workflow management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "a-sdlc contributors" }
]
keywords = [
    "sdlc",
    "claude-code",
    "documentation",
    "automation",
    "prd",
    "requirements",
    "tasks"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
linear = [
    "httpx>=0.25.0",
]

[project.scripts]
a-sdlc = "a_sdlc.cli:main"

[project.entry-points."a_sdlc.plugins"]
local = "a_sdlc.plugins.local:LocalPlugin"
linear = "a_sdlc.plugins.linear:LinearPlugin"

[project.urls]
Homepage = "https://github.com/a-sdlc/a-sdlc"
Documentation = "https://github.com/a-sdlc/a-sdlc#readme"
Repository = "https://github.com/a-sdlc/a-sdlc"
Issues = "https://github.com/a-sdlc/a-sdlc/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/a_sdlc"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
